refactor(cm): carry a canonical intrinsic by identity, not by its rendered name
A `canon future.read` is typed — the Component Model instantiates one per
`future<T>` — so the core module needs a distinct import name per payload. That
much the CM forces. What did not follow is using the name as the carrier:
synthesis classified the payload, rendered it into a `CmRawCall`'s string, and
WIR parsed it back with `from_import_name`. The round-trip is what made the
encoding have to be invertible, and it is not: the HTTP trailers future renders
to the bare base name, which is also what a payload nobody classified produced.
A `future<Color>` therefore came back as the trailers future and ICEd in
codegen asking the import plan for an HTTP interface.

Carry the identity instead. `CmRawCall` now holds a `CmCallTarget` — a WASI
alias or a `CanonicalIntrinsic` — through TIR and NIR into WIR, where the
intrinsic is used directly. `import_name` becomes a terminal rendering, so it
only has to be injective, and the names in the emitted module are unchanged.

Move the intrinsic and payload types out of `wir.rs` into `canonical.rs`: they
describe a Component Model concept, and TIR has to name them now that it
carries one.

Typing the target found four more sites that were fabricating names with
`format!` or a literal — `task-return`, `task-return:<export>`,
`resource-drop:<cm>`, and `stream-read:<record>`. Each now names its intrinsic,
so the rendering lives in one place.

Stop `from_import_name` inventing a payload for a suffix-less future name.
`#[canonical("wasi", "future-read")]` is a template the call site
parameterizes — its own doc comment says so — and reading it as the trailers
future is what the `future_payload().is_none()` guard in import registration
was working around. The guard is gone with the cause.
